implement clean for model v2
